move `ContractTransactionChecks` to `ledger-api-integration-tests`
